MD5: Fix partial key generation errors
When verifying a key, the generation method sometimes returns
a wrong hash. This behaviour does not occur when cleaning up the
sensitive data at the end of the method.

It seems that there are some references to 'A', 'Alternate' or
'AlternateSum' which will be used simultaneously.

This fix doesn't solve the cause, it only runs the cleanup at
the end of the method.